+++ src/PyR3/contrib/factories/SCurve.py
@@ -3,10 +3,13 @@ from __future__ import annotations
 
 import numpy
 
+from PyR3.factory.fields.BSDF_Material import BSDF_Material
 from PyR3.factory.fields.Number import Integer
 from PyR3.factory.fields.Unit import Length
 from PyR3.factory.MeshFactory import MeshFactory
+from PyR3.shortcut.context import Objects
 from PyR3.shortcut.edit import Edit
+from PyR3.shortcut.material import apply_BSDF_material_params
 from PyR3.shortcut.mesh import continuous_edge, fromPyData
 from PyR3.shortcut.modifiers import Solidify
 from PyR3.shortcut.transform import Transform
@@ -25,6 +28,7 @@ class SCurve(MeshFactory):
     upper_length = Length()
     steps = Integer(value_range=range(4, 128))
     thickness = Length()
+    material = BSDF_Material()
 
     def render(self) -> None:
         half_total_width = self.total_width / 2
@@ -38,6 +42,7 @@ class SCurve(MeshFactory):
             Transform.move((0, self.total_width, 0))
         # final touches - thickness and volume quality
         Solidify(curve_mesh, self.thickness, offset=0).apply()
+        apply_BSDF_material_params(Objects.active, self.material)
 
     def generate_vertices_list(self, half_total_width, half_total_height):
         lower_limit = -half_total_width + self.lower_length