+++ CHANGELOG.rst
@@ -61,5 +61,11 @@ Changelog
 * Updated implementation of PyR3.contrib.factories.CapacitorCase
 * Added API and CLI for rendering single models using MeshFactories
 * Added Remesh modifier class
+<<<<<<< Updated upstream
 * Added SCurve MeshFactory subclass
 * Added BSDF_Material, Color, HeterotypeSequence and HomotypeSequence Field types
+=======
+* Added CapacitorCase and SCurve MeshFactory subclasses
+* Added BSDF_Material, Color, HeterotypeSequence and HomotypeSequence Field types
+* Added PyR3.shortcut.material.apply_BSDF_material_params() function for high level applying of material params
+>>>>>>> Stashed changes